made a new class commands; to sort out commands from the args reader.
