+++ Src/BootCamp.Chapter/ArgsReader.cs
@@ -1,20 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Text;
 
 namespace BootCamp.Chapter
 {
     public static class ArgsReader
     {
-        private enum ArgsNumber
-        {
-            FileToRead,
-            command,
-            FileToWrite
-        }
-
         //TODO finish ArgsReader
         /*
         You need to generate the following reports:
@@ -34,26 +24,29 @@ namespace BootCamp.Chapter
         Sorting is just 1 extra arg: -asc (sorts in ascending order by shop name); -desc (sorts in descending order by shop name)
         */
 
-        const string timeCommand = "time";
-        const string cityCommand = "city";
+        private const string timeCommand = "time";
+        private const string cityCommand = "city";
+        const int fileToRead = 0;
+        const int command = 1;
+        const int fileToWrite = 2;
         public static void Read(string[] args)
         {
             ArgsChecksLength(args);
-            List<Transaction> transactions = ReportsManager.ReadTransaction(args[(int)ArgsNumber.FileToRead]);
-            string[] command = ReadCommand(args[(int)ArgsNumber.command]);
+            List<Transaction> transactions = ReportsManager.ReadTransaction(args[fileToRead]);
+            string[] commandArr = ReadCommand(args[command]);
             IEnumerable<String> toBeWritten = default;
 
-            switch (command[0])
+            switch (commandArr[0])
             {
                 case timeCommand:
-                    toBeWritten = RunTimeCommand(command, transactions);
+                    toBeWritten = Command.RunTimeCommand(commandArr, transactions);
                     break;
                 case cityCommand:
-                    toBeWritten = RunCityCommand(command, transactions);
+                    toBeWritten = Command.RunCityCommand(commandArr, transactions);
                     break;
             }
 
-            ReportsManager.WriteTransaction(args[2], toBeWritten);
+            ReportsManager.WriteTransaction(args[fileToWrite], toBeWritten);
         }
 
         private static string[] ReadCommand(string command)
@@ -65,121 +58,24 @@ namespace BootCamp.Chapter
 
             string[] splitCommand = command.Split(' ');
 
-                //Trim all the Strings
-                for (int i = 0; i < splitCommand.Length; i++)
-                {
+            //Trim all the Strings
+            for (int i = 0; i < splitCommand.Length; i++)
+            {
                 splitCommand[i] = splitCommand[i].Trim();
-                }
+            }
 
             switch (splitCommand[0])
             {
                 case timeCommand:
-                    return(splitCommand);
+                    return (splitCommand);
                 case cityCommand:
-                    return(splitCommand);
+                    return (splitCommand);
                 default:
                     throw new InvalidCommandException($"{splitCommand[0]} is not a valid command.");
             }
 
         }
 
-        private static IEnumerable<String> RunTimeCommand(string[] command, List<Transaction> transactions)
-        {
-            //TODO RunTimeCommand
-            /*
-            By time(time)
-            how many items have been bought during every hour of time of day,
-            how much money did every hour total(on average),
-            and get rush hour(most mony earned on average).
-            Support getting items sold count and money earned for a selected range of hours as well.
-            */
-            //"time 11:00-17:00"
-            if (command.Length == 1)
-            {
-                return Time(transactions);
-            }
-            else if (command.Length == 2)
-            {
-                if(IsHoursCorrect(command[1], out DateTime[] times))
-                {
-                   return Time(transactions, times);
-                }
-                throw new InvalidCommandException($"{command[0]} has the wrong Times.");
-            }
-            else
-            {
-                throw new InvalidCommandException($"{command[0]} has to many parameters.");
-            }
-
-        }
-
-        private static bool IsHoursCorrect(string hours, out DateTime[] times)
-        {
-            //TODO IsHoursCorrect()
-            throw new NotImplementedException();
-        }
-
-        private static IEnumerable<String> Time(List<Transaction> transactions)
-        {
-            const string topRow = "Hour, Count, Earned";
-
-            var soldByTime = transactions.GroupBy(t => t.DateTime.Hour).Select(z => new TimeNumberEarned
-            {
-
-               Time = z.First().DateTime.Hour,
-               Number = z.Count(),
-               Earned =  z.Sum(x => x.Price) / z.Select(x => x.DateTime.Date).Distinct().Count()
-            }
-            ).ToList();
-
-            return CreateTabel(topRow, soldByTime, 0, 24);
-        }
-        private static IEnumerable<String> CreateTabel(string topRow, IEnumerable<TimeNumberEarned> soldByTime, int startTime, int EndTime)
-        {
-            List<String> toBeWritten = new List<string>();
-
-            //00, 0, "0,00 €"
-            toBeWritten.Add(topRow);
-            int rushHour = 0;
-            decimal mostEarned = Decimal.MinValue;
-            for (int i = startTime; i < EndTime; i++)
-            {
-                int count = 0;
-                decimal earned = 0;
-
-                foreach (TimeNumberEarned timeNumberEarned in soldByTime)
-                {
-                    if(timeNumberEarned.Time == i)
-                    {
-                        
-                        count = timeNumberEarned.Number;
-                        earned = timeNumberEarned.Earned;
-                        if (mostEarned < earned)
-                        {
-                            mostEarned = earned;
-                            rushHour = i;
-                        }
-                    }
-                }
-                toBeWritten.Add($"{i.ToString("D2")}, {count}, \"{earned.ToString("C2", CultureInfo.CurrentCulture)}\"");
-            }
-
-            //Rush hour: 22
-            toBeWritten.Add($"Rush hour: {rushHour}");
-
-            return toBeWritten;
-        }
-        private static IEnumerable<String> Time(List<Transaction> transactions, DateTime[] times)
-        {
-            //TODO Time With Time
-            throw new NotImplementedException();
-        }
-        private static IEnumerable<String> RunCityCommand(string[] command, List<Transaction> transactions)
-        {
-            //TODO RunCityCommand
-            //What city(can be parsed from address) earned the most / least money and what city sold the most / least items ? (city[-min / -max][-items / -money])
-            throw new NotImplementedException();
-        }
         private static void ArgsChecksLength(string[] args)
         {
             if (args == null || args.Length != 3)
